frontend/requirements.txt
Flask==2.3.2
gunicorn==22.0.0
requests==2.31.0
Jinja2==3.1.4
Pillow==10.3.0
Werkzeug==3.0.3
python-dotenv==1.0.0
google-generativeai==0.7.2
